Update secrets encryption algorithm information
I think the description 'are encrypted using AES-256' is likely incorrect. AES is a symmetric encryption algorithm, which would mean the same key would be used to encrypt the credentials in the tenant and decrypt them in the AU. In other words, the tenant would have access to the key material to decrypt the customer secrets. I suspect that the actual design is likely using an asymmetric encryption algorithm, with the secrets encrypted by the public key prior to storing them on the tenant and the private key (the only way to decrypt the secrets) residing only on the AU?